Move ICustomTypeMarshallingStrategy implementations into separate files
