+++ src/libraries/System.Runtime.InteropServices/gen/Microsoft.Interop.SourceGeneration/Marshalling/StatefulMarshallingStrategy.cs
@@ -1,0 +1,310 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
+
+namespace Microsoft.Interop
+{
+    internal sealed class StatefulValueMarshalling : ICustomTypeMarshallingStrategy
+    {
+        internal const string MarshallerIdentifier = "marshaller";
+        private readonly TypeSyntax _marshallerTypeSyntax;
+        private readonly TypeSyntax _nativeTypeSyntax;
+        private readonly MarshallerShape _shape;
+
+        public StatefulValueMarshalling(TypeSyntax marshallerTypeSyntax, TypeSyntax nativeTypeSyntax, MarshallerShape shape)
+        {
+            _marshallerTypeSyntax = marshallerTypeSyntax;
+            _nativeTypeSyntax = nativeTypeSyntax;
+            _shape = shape;
+        }
+
+        public TypeSyntax AsNativeType(TypePositionInfo info)
+        {
+            return _nativeTypeSyntax;
+        }
+
+        public bool UsesNativeIdentifier(TypePositionInfo info, StubCodeContext context) => true;
+
+        public IEnumerable<StatementSyntax> GenerateCleanupStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            if (!_shape.HasFlag(MarshallerShape.Free))
+                yield break;
+
+            // <marshaller>.Free();
+            yield return ExpressionStatement(
+                InvocationExpression(
+                    MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                        IdentifierName(context.GetAdditionalIdentifier(info, MarshallerIdentifier)),
+                        IdentifierName(ShapeMemberNames.Free)),
+                    ArgumentList()));
+        }
+
+        public IEnumerable<StatementSyntax> GenerateGuaranteedUnmarshalStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            if (!_shape.HasFlag(MarshallerShape.GuaranteedUnmarshal))
+                yield break;
+
+            (string managedIdentifier, _) = context.GetIdentifiers(info);
+
+            // <managedIdentifier> = <marshaller>.ToManagedFinally();
+            yield return ExpressionStatement(
+                AssignmentExpression(
+                    SyntaxKind.SimpleAssignmentExpression,
+                    IdentifierName(managedIdentifier),
+                    InvocationExpression(
+                        MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                            IdentifierName(context.GetAdditionalIdentifier(info, MarshallerIdentifier)),
+                            IdentifierName(ShapeMemberNames.Value.Stateful.ToManagedFinally)),
+                        ArgumentList())));
+        }
+
+        public IEnumerable<StatementSyntax> GenerateMarshalStatements(TypePositionInfo info, StubCodeContext context, IEnumerable<ArgumentSyntax> nativeTypeConstructorArguments)
+        {
+            if (!_shape.HasFlag(MarshallerShape.ToUnmanaged))
+                yield break;
+
+            (string managedIdentifier, _) = context.GetIdentifiers(info);
+
+            // <marshaller>.FromManaged(<managedIdentifier>);
+            yield return ExpressionStatement(
+                InvocationExpression(
+                    MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                        IdentifierName(context.GetAdditionalIdentifier(info, MarshallerIdentifier)),
+                        IdentifierName(ShapeMemberNames.Value.Stateful.FromManaged)),
+                    ArgumentList(SingletonSeparatedList(
+                        Argument(IdentifierName(managedIdentifier))))));
+        }
+
+        public IEnumerable<StatementSyntax> GeneratePinnedMarshalStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            if (!_shape.HasFlag(MarshallerShape.ToUnmanaged) && !_shape.HasFlag(MarshallerShape.CallerAllocatedBuffer))
+                yield break;
+
+            (_, string nativeIdentifier) = context.GetIdentifiers(info);
+
+            // <nativeIdentifier> = <marshaller>.ToUnmanaged();
+            yield return ExpressionStatement(
+                AssignmentExpression(
+                    SyntaxKind.SimpleAssignmentExpression,
+                    IdentifierName(nativeIdentifier),
+                    InvocationExpression(
+                        MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                            IdentifierName(context.GetAdditionalIdentifier(info, MarshallerIdentifier)),
+                            IdentifierName(ShapeMemberNames.Value.Stateful.ToUnmanaged)),
+                        ArgumentList())));
+        }
+
+        public IEnumerable<StatementSyntax> GenerateUnmarshalStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            if (!_shape.HasFlag(MarshallerShape.ToManaged))
+                yield break;
+
+            (string managedIdentifier, _) = context.GetIdentifiers(info);
+
+            // <managedIdentifier> = <marshaller>.ToManaged();
+            yield return ExpressionStatement(
+                AssignmentExpression(
+                    SyntaxKind.SimpleAssignmentExpression,
+                    IdentifierName(managedIdentifier),
+                    InvocationExpression(
+                        MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                            IdentifierName(context.GetAdditionalIdentifier(info, MarshallerIdentifier)),
+                            IdentifierName(ShapeMemberNames.Value.Stateful.ToManaged)),
+                        ArgumentList())));
+        }
+
+        public IEnumerable<StatementSyntax> GenerateUnmarshalCaptureStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            if (!_shape.HasFlag(MarshallerShape.ToManaged) && !_shape.HasFlag(MarshallerShape.GuaranteedUnmarshal))
+                yield break;
+
+            (_, string nativeIdentifier) = context.GetIdentifiers(info);
+
+            // <marshaller>.FromUnmanaged(<nativeIdentifier>);
+            yield return ExpressionStatement(
+                InvocationExpression(
+                    MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                        IdentifierName(context.GetAdditionalIdentifier(info, MarshallerIdentifier)),
+                        IdentifierName(ShapeMemberNames.Value.Stateful.FromUnmanaged)),
+                    ArgumentList(SingletonSeparatedList(
+                        Argument(IdentifierName(nativeIdentifier))))));
+        }
+
+        public IEnumerable<ArgumentSyntax> GetNativeTypeConstructorArguments(TypePositionInfo info, StubCodeContext context)
+        {
+            return Array.Empty<ArgumentSyntax>();
+        }
+
+        public IEnumerable<StatementSyntax> GenerateSetupStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            yield return MarshallerHelpers.Declare(
+                _marshallerTypeSyntax,
+                context.GetAdditionalIdentifier(info, MarshallerIdentifier),
+                ImplicitObjectCreationExpression(ArgumentList(), initializer: null));
+        }
+
+        public IEnumerable<StatementSyntax> GeneratePinStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            if (!_shape.HasFlag(MarshallerShape.StatefulPinnableReference))
+                yield break;
+
+            string unusedIdentifier = context.GetAdditionalIdentifier(info, "unused");
+            yield return FixedStatement(
+                VariableDeclaration(
+                    PointerType(PredefinedType(Token(SyntaxKind.VoidKeyword))),
+                    SingletonSeparatedList(
+                        VariableDeclarator(unusedIdentifier)
+                            .WithInitializer(EqualsValueClause(IdentifierName(context.GetAdditionalIdentifier(info, MarshallerIdentifier)))))),
+                EmptyStatement());
+        }
+
+        public IEnumerable<StatementSyntax> GenerateNotifyForSuccessfulInvokeStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            if (!_shape.HasFlag(MarshallerShape.OnInvoked))
+                yield break;
+
+            // <marshaller>.OnInvoked();
+            yield return ExpressionStatement(
+                InvocationExpression(
+                    MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                        IdentifierName(context.GetAdditionalIdentifier(info, MarshallerIdentifier)),
+                        IdentifierName(ShapeMemberNames.Value.Stateful.OnInvoked)),
+                    ArgumentList()));
+        }
+    }
+
+    /// <summary>
+    /// Marshaller that enables support for a stackalloc constructor variant on a native type.
+    /// </summary>
+    internal sealed class StatefulCallerAllocatedBufferMarshalling : ICustomTypeMarshallingStrategy
+    {
+        private readonly ICustomTypeMarshallingStrategy _innerMarshaller;
+        private readonly TypeSyntax _marshallerType;
+        private readonly TypeSyntax _bufferElementType;
+
+        public StatefulCallerAllocatedBufferMarshalling(ICustomTypeMarshallingStrategy innerMarshaller, TypeSyntax marshallerType, TypeSyntax bufferElementType)
+        {
+            _innerMarshaller = innerMarshaller;
+            _marshallerType = marshallerType;
+            _bufferElementType = bufferElementType;
+        }
+
+        public TypeSyntax AsNativeType(TypePositionInfo info)
+        {
+            return _innerMarshaller.AsNativeType(info);
+        }
+
+        public IEnumerable<StatementSyntax> GenerateCleanupStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            return _innerMarshaller.GenerateCleanupStatements(info, context);
+        }
+
+        public IEnumerable<StatementSyntax> GenerateMarshalStatements(TypePositionInfo info, StubCodeContext context, IEnumerable<ArgumentSyntax> nativeTypeConstructorArguments)
+        {
+            if (CanUseCallerAllocatedBuffer(info, context))
+            {
+                return GenerateCallerAllocatedBufferMarshalStatements();
+            }
+
+            return _innerMarshaller.GenerateMarshalStatements(info, context, nativeTypeConstructorArguments);
+
+            IEnumerable<StatementSyntax> GenerateCallerAllocatedBufferMarshalStatements()
+            {
+                // TODO: Update once we can consume the scoped keword. We should be able to simplify this once we get that API.
+                string stackPtrIdentifier = context.GetAdditionalIdentifier(info, "stackptr");
+                // <bufferElementType>* <managedIdentifier>__stackptr = stackalloc <bufferElementType>[<_bufferSize>];
+                yield return LocalDeclarationStatement(
+                VariableDeclaration(
+                    PointerType(_bufferElementType),
+                    SingletonSeparatedList(
+                        VariableDeclarator(stackPtrIdentifier)
+                            .WithInitializer(EqualsValueClause(
+                                StackAllocArrayCreationExpression(
+                                        ArrayType(
+                                            _bufferElementType,
+                                            SingletonList(ArrayRankSpecifier(SingletonSeparatedList<ExpressionSyntax>(
+                                                MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                                                    _marshallerType,
+                                                    IdentifierName(ShapeMemberNames.BufferSize))
+                                            ))))))))));
+
+
+                (string managedIdentifier, _) = context.GetIdentifiers(info);
+
+                // <marshaller>.FromManaged(<managedIdentifier>);
+                yield return ExpressionStatement(
+                    InvocationExpression(
+                        MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                            IdentifierName(context.GetAdditionalIdentifier(info, StatefulValueMarshalling.MarshallerIdentifier)),
+                            IdentifierName(ShapeMemberNames.Value.Stateful.FromManaged)),
+                        ArgumentList(SeparatedList(
+                            new[]
+                            {
+                                Argument(IdentifierName(managedIdentifier)),
+                                Argument(
+                                    ObjectCreationExpression(
+                                        GenericName(Identifier(TypeNames.System_Span),
+                                            TypeArgumentList(SingletonSeparatedList(
+                                                _bufferElementType))))
+                                    .WithArgumentList(
+                                        ArgumentList(SeparatedList(new ArgumentSyntax[]
+                                        {
+                                            Argument(IdentifierName(stackPtrIdentifier)),
+                                            Argument(MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                                                    _marshallerType,
+                                                    IdentifierName(ShapeMemberNames.BufferSize)))
+                                        }))))
+                            }))));
+            }
+        }
+
+        public IEnumerable<StatementSyntax> GeneratePinnedMarshalStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            return _innerMarshaller.GeneratePinnedMarshalStatements(info, context);
+        }
+
+        private static bool CanUseCallerAllocatedBuffer(TypePositionInfo info, StubCodeContext context)
+        {
+            return context.SingleFrameSpansNativeContext && (!info.IsByRef || info.RefKind == RefKind.In);
+        }
+
+        public IEnumerable<StatementSyntax> GeneratePinStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            return _innerMarshaller.GeneratePinStatements(info, context);
+        }
+
+        public IEnumerable<StatementSyntax> GenerateSetupStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            return _innerMarshaller.GenerateSetupStatements(info, context);
+        }
+
+        public IEnumerable<StatementSyntax> GenerateUnmarshalCaptureStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            return _innerMarshaller.GenerateUnmarshalCaptureStatements(info, context);
+        }
+
+        public IEnumerable<StatementSyntax> GenerateUnmarshalStatements(TypePositionInfo info, StubCodeContext context)
+        {
+            return _innerMarshaller.GenerateUnmarshalStatements(info, context);
+        }
+
+        public IEnumerable<ArgumentSyntax> GetNativeTypeConstructorArguments(TypePositionInfo info, StubCodeContext context)
+        {
+            return Array.Empty<ArgumentSyntax>();
+        }
+
+        public bool UsesNativeIdentifier(TypePositionInfo info, StubCodeContext context)
+        {
+            return _innerMarshaller.UsesNativeIdentifier(info, context);
+        }
+
+        public IEnumerable<StatementSyntax> GenerateGuaranteedUnmarshalStatements(TypePositionInfo info, StubCodeContext context) => _innerMarshaller.GenerateGuaranteedUnmarshalStatements(info, context);
+        public IEnumerable<StatementSyntax> GenerateNotifyForSuccessfulInvokeStatements(TypePositionInfo info, StubCodeContext context) => _innerMarshaller.GenerateNotifyForSuccessfulInvokeStatements(info, context);
+    }
+}
